+++ common/storybook_utils.js
@@ -1,5 +1,4 @@
-import * as dialtoneIcons from '@dialpad/dialtone-icons';
-import { pascalCaseToKebabCase } from '@/common/utils';
+import iconNames from '@dialpad/dialtone-icons/dist/icons.json';
 
 /**
  * Will use a Vue SFC to render the template rather than a template string.
@@ -13,18 +12,18 @@ import { pascalCaseToKebabCase } from '@/common/utils';
  *                                      but rather the usage of that component.
  * @returns {component} the template component with props and args added.
  */
-export const createTemplateFromVueFile = (args, argTypes, templateComponent) => {
-  templateComponent.props = Object.keys(argTypes);
-  templateComponent.args = args;
-  return templateComponent;
-};
+export const createTemplateFromVueFile = (args, argTypes, templateComponent) => ({
+  components: { templateComponent },
+  props: Object.keys(argTypes),
+  template: '<template-component v-bind="$props"></template-component>',
+});
 
 /**
  * Gets the full list of icon component names from the dialtone package
  * @returns {string[]} icon component names
  */
 export function getIconNames () {
-  return [undefined, ...Object.keys(dialtoneIcons).map(name => pascalCaseToKebabCase(name))];
+  return [undefined, ...iconNames];
 }
 
 export const generateTemplate = (component,