+++ components/icon/icon.vue
@@ -1,7 +1,8 @@
 <template>
   <component
-    :is="currentIcon"
-    v-if="currentIcon"
+    :is="icon"
+    v-if="icon"
+    :id="id"
     data-qa="dt-icon"
     :aria-hidden="ariaLabel ? 'false' : 'true'"
     :aria-label="ariaLabel"
@@ -10,8 +11,13 @@
 </template>
 
 <script>
-import { ICON_SIZE_MODIFIERS, DIALTONE_ICONS } from './icon_constants';
-import { kebabCaseToPascalCase } from '@/common/utils';
+import { ICON_SIZE_MODIFIERS } from './icon_constants';
+import { getUniqueString } from '@/common/utils.js';
+
+const dialtoneIcons = import.meta.glob(
+  '/node_modules/@dialpad/dialtone-icons/dist/svg/*.svg',
+  { as: 'component', eager: true },
+);
 
 /**
  * The Icon component provides a set of glyphs and sizes to provide context your application.
@@ -21,6 +27,16 @@ export default {
   name: 'DtIcon',
 
   props: {
+    /**
+     * DtIcon identifier
+     */
+    id: {
+      type: String,
+      default () {
+        return getUniqueString();
+      },
+    },
+
     /**
      * The size of the icon.
      * @values 100, 200, 300, 400, 500, 600, 700, 800
@@ -53,12 +69,9 @@ export default {
       return ICON_SIZE_MODIFIERS[this.size];
     },
 
-    iconName () {
-      return kebabCaseToPascalCase(this.name);
-    },
-
-    currentIcon () {
-      return DIALTONE_ICONS[this.iconName];
+    icon () {
+      const iconPath = `/node_modules/@dialpad/dialtone-icons/dist/svg/${this.name}.svg`;
+      return dialtoneIcons[iconPath].default;
     },
   },
 };