+++ components/tooltip/tooltip.stories.js
@@ -3,6 +3,7 @@ import DtTooltip from './tooltip.vue';
 import DtTooltipFlipTemplate from './tooltip_flip.story.vue';
 import DtTooltipDefault from './tooltip_default.story.vue';
 import DtTooltipVariantsTemplate from './tooltip_variants.vue';
+import DtTooltipChangeOnClick from './tooltip_change_on_click.story.vue';
 import { action } from '@storybook/addon-actions';
 
 import { TOOLTIP_DIRECTIONS, TOOLTIP_STICKY_VALUES } from './tooltip_constants';
@@ -14,6 +15,7 @@ export const argsData = {
   default: `This is a simple tooltip. You can set the position of the tooltip using the placement prop!`,
   sticky: false,
   onShown: action('shown'),
+  showTooltip: null,
 };
 
 export const argTypesData = {
@@ -123,6 +125,8 @@ const TooltipDefaultTemplate = (args, { argTypes }) =>
   createTemplateFromVueFile(args, argTypes, DtTooltipDefault);
 const TooltipVariantsTemplate = (args, { argTypes }) =>
   createTemplateFromVueFile(args, argTypes, DtTooltipVariantsTemplate);
+const TooltipChangeOnClick = (args, { argTypes }) =>
+  createTemplateFromVueFile(args, argTypes, DtTooltipChangeOnClick);
 
 export const Default = {
   render: TooltipDefaultTemplate,
@@ -135,6 +139,11 @@ export const Variants = {
   parameters: {
     options: { showPanel: false },
     controls: { disable: true },
+    percy: {
+      args: {
+        showTooltip: true,
+      },
+    },
   },
 };
 
@@ -148,5 +157,13 @@ export const Flip = {
   parameters: {
     options: { showPanel: false },
     controls: { disable: true },
+  },
+};
+
+export const ChangeOnClick = {
+  render: TooltipChangeOnClick,
+  args: {
+    anchor: 'Click to see the tooltip content change',
+    sticky: 'popper',
   },
 };