+++ recipes/comboboxes/combobox_multi_select/combobox_multi_select_default.story.vue
@@ -1,31 +1,31 @@
 <template>
   <dt-recipe-combobox-multi-select
     ref="comboboxMultiSelect"
-    :label="label"
-    :label-visible="labelVisible"
-    :description="description"
-    :size="size"
-    :loading="loading"
-    :loading-message="loadingMessage"
-    :show-list="showList"
-    :selected-items="selectedItems"
-    :max-selected="maxSelected"
-    :list-max-height="listMaxHeight"
-    :max-selected-message="maxSelectedMessage"
-    :has-suggestion-list="hasSuggestionList"
-    :visually-hidden-close-label="visuallyHiddenCloseLabel"
-    :append-to="appendTo"
-    :transition="transition"
+    :label="$attrs.label"
+    :label-visible="$attrs.labelVisible"
+    :description="$attrs.description"
+    :size="$attrs.size"
+    :loading="$attrs.loading"
+    :loading-message="$attrs.loadingMessage"
+    :show-list="$attrs.showList"
+    :selected-items="$attrs.selectedItems"
+    :max-selected="$attrs.maxSelected"
+    :list-max-height="$attrs.listMaxHeight"
+    :max-selected-message="$attrs.maxSelectedMessage"
+    :has-suggestion-list="$attrs.hasSuggestionList"
+    :visually-hidden-close-label="$attrs.visuallyHiddenCloseLabel"
+    :append-to="$attrs.appendTo"
+    :transition="$attrs.transition"
     @input="onComboboxInput"
     @select="onComboboxSelect"
     @remove="onComboboxRemove"
-    @max-selected="onComboboxMaxSelected"
+    @max-selected="$attrs.onMaxSelected"
   >
     <template
-      v-if="header"
+      v-if="$attrs.header"
       #header
     >
-      <span v-html="header" />
+      <span v-html="$attrs.header" />
     </template>
     <template #list>
       <ul
@@ -46,10 +46,10 @@
       </ul>
     </template>
     <template
-      v-if="footer"
+      v-if="$attrs.footer"
       #footer
     >
-      <span v-html="footer" />
+      <span v-html="$attrs.footer" />
     </template>
   </dt-recipe-combobox-multi-select>
 </template>
@@ -76,18 +76,18 @@ export default {
     onComboboxInput (value) {
       // Filter list
       this.items = ITEMS_LIST_DATA.filter(item => item.value.includes(value));
-      this.onInput(value);
+      this.$attrs.onInput(value);
     },
 
     onComboboxSelect (i) {
       if (this.items[i]) {
-        this.onSelect(i);
+        this.$attrs.onSelect(i);
 
         const item = this.items[i].value;
-        if (this.selectedItems.includes(item)) {
+        if (this.$attrs.selectedItems.includes(item)) {
           return;
         }
-        this.selectedItems.push(item);
+        this.$attrs.selectedItems.push(item);
 
         // Clear input box and unfilter list
         this.$refs.comboboxMultiSelect.$data.value = '';
@@ -95,17 +95,13 @@ export default {
       }
     },
 
-    onComboboxMaxSelected () {
-      this.onMaxSelected();
-    },
-
     onComboboxRemove (item) {
-      this.onRemove(item);
-      const index = this.selectedItems.indexOf(item);
+      this.$attrs.onRemove(item);
+      const index = this.$attrs.selectedItems.indexOf(item);
       if (index < 0) {
         return;
       }
-      this.selectedItems.splice(index, 1);
+      this.$attrs.selectedItems.splice(index, 1);
     },
   },
 };