+++ vite.config.js
@@ -1,7 +1,9 @@
 import { defineConfig } from 'vite';
+import { createSvgPlugin } from 'vite-plugin-vue2-svg';
 import vue from '@vitejs/plugin-vue2';
 import path, { resolve } from 'path';
 import { fileURLToPath } from 'url';
+import hash from 'string-hash';
 
 const __filename = fileURLToPath(import.meta.url);
 const __dirname = path.dirname(__filename);
@@ -17,7 +19,10 @@ export default defineConfig({
       },
     },
     rollupOptions: {
-      external: ['vue', '@dialpad/dialtone'],
+      external: [
+        'vue',
+        '@dialpad/dialtone',
+      ],
       output: {
         globals: {
           vue: 'Vue',
@@ -25,7 +30,22 @@ export default defineConfig({
       },
     },
   },
-  plugins: [vue()],
+  plugins: [vue(), createSvgPlugin({
+    svgoConfig: {
+      plugins: [
+        {
+          name: 'prefixIds',
+          params: {
+            delim: '',
+            prefix: (_, extra) => {
+              return `dt-icon${hash(extra.path)}`;
+            },
+            prefixClassNames: false,
+          },
+        },
+      ],
+    },
+  })],
   resolve: {
     alias: {
       '@': fileURLToPath(new URL('.', import.meta.url)),